Mcp_Server_OpenAI/requirements.txt
fastmcp
pymongo
openai
python-dotenv
